+++ src/AtomUI.Desktop.Controls/Window/Chrome/MacOSWindowChromeManager.cs
@@ -24,6 +24,8 @@ internal sealed class MacOSWindowChromeManager : IWindowChromeManager
 
     public bool UsesCustomResizer => false;
 
+    public bool SupportsPinCaptionButton => true;
+
     public Action? PrepareInitialShowState()
     {
         if (_initialShowStatePrepared || _window.IsVisible)