+++ src/AtomUI.Desktop.Controls/WindowTitleBar/CaptionButtonAction.cs
@@ -1,0 +1,10 @@
+namespace AtomUI.Desktop.Controls;
+
+internal enum CaptionButtonAction
+{
+    Minimize,
+    ToggleMaximize,
+    ToggleFullScreen,
+    TogglePin,
+    Close
+}